+++ packages/core/src/rule-processor.ts
@@ -1,6 +1,6 @@
 import type { CssOptions, Stylesheet } from '@pandacss/core'
 import type { RecipeDefinition, SlotRecipeDefinition, SystemStyleObject } from '@pandacss/types'
-import type { CoreContext } from './core-context'
+import type { Context } from './context'
 import type { StyleDecoder } from './style-decoder'
 import type { StyleEncoder } from './style-encoder'
 
@@ -10,7 +10,7 @@ export class RuleProcessor {
   sheet: Stylesheet | undefined
   params: Pick<PrepareParams, 'encoder' | 'decoder'>
 
-  constructor(private context: CoreContext, params?: Pick<PrepareParams, 'encoder' | 'decoder'>) {
+  constructor(private context: Context, params?: Pick<PrepareParams, 'encoder' | 'decoder'>) {
     this.params = params ?? {
       encoder: context.encoder,
       decoder: context.decoder,
@@ -47,7 +47,6 @@ export class RuleProcessor {
 
   css(styles: SystemStyleObject): AtomicRule {
     const { encoder, decoder, sheet } = this.prepare()
-
     encoder.processAtomic(styles)
     decoder.collect(encoder)
 
@@ -61,15 +60,9 @@ export class RuleProcessor {
     }
   }
 
-  cva(recipeConfig: RecipeDefinition<any> | SlotRecipeDefinition<string, any>): AtomicRecipeRule {
+  cva(recipeConfig: RecipeDefinition<any>): AtomicRecipeRule {
     const { encoder, decoder, sheet } = this.prepare()
-
-    if ('slots' in recipeConfig) {
-      encoder.processAtomicSlotRecipe(recipeConfig)
-    } else {
-      encoder.processAtomicRecipe(recipeConfig)
-    }
-
+    encoder.processAtomicRecipe(recipeConfig)
     decoder.collect(encoder)
 
     return {
@@ -83,18 +76,23 @@ export class RuleProcessor {
   }
 
   sva(recipeConfig: SlotRecipeDefinition<string, any>): AtomicRecipeRule {
-    return this.cva(recipeConfig)
+    const { encoder, decoder, sheet } = this.prepare()
+    encoder.processAtomicSlotRecipe(recipeConfig)
+    decoder.collect(encoder)
+    return {
+      config: recipeConfig,
+      className: Array.from(decoder.classNames.keys()),
+      toCss: (options?: CssOptions) => {
+        sheet.processDecoder(decoder)
+        return sheet.toCss({ optimize: true, ...options })
+      },
+    }
   }
 
   recipe(name: string, variants: Record<string, any>): RecipeRule | undefined {
-    const recipeConfig = this.context.recipes.getConfig(name)
-    if (!recipeConfig) return
-
     const { encoder, decoder, sheet } = this.prepare()
-
     encoder.processRecipe(name, variants)
     decoder.collect(encoder)
-
     return {
       variants,
       className: Array.from(decoder.classNames.keys()),