+++ packages/core/src/stylesheet.ts
@@ -1,8 +1,9 @@
 import { logger } from '@pandacss/logger'
-import type { CascadeLayer, Dict, StyleDecoderInterface, SystemStyleObject } from '@pandacss/types'
+import type { CascadeLayer, Dict, SystemStyleObject } from '@pandacss/types'
 import postcss, { CssSyntaxError } from 'postcss'
 import { expandCssFunctions, optimizeCss } from './optimize'
 import { serializeStyles } from './serialize'
+import type { StyleDecoder } from './style-decoder'
 import { toCss } from './to-css'
 import type { CssOptions, LayerName, ProcessOptions, StylesheetContext } from './types'
 
@@ -30,7 +31,7 @@ export class Stylesheet {
       layer.append(toCss(styles).toString())
     } catch (error) {
       if (error instanceof CssSyntaxError) {
-        logger.error('sheet:process', error)
+        logger.error('sheet:process', error.showSourceCode(true))
       }
     }
     return
@@ -47,7 +48,7 @@ export class Stylesheet {
     this.process({ styles, layer })
   }
 
-  processDecoder = (decoder: StyleDecoderInterface) => {
+  processDecoder = (decoder: StyleDecoder) => {
     decoder.atomic.forEach((css) => {
       this.processCss(css.result, (css.layer as LayerName) ?? 'utilities')
     })
@@ -90,13 +91,7 @@ export class Stylesheet {
       return optimize ? optimizeCss(css, { minify }) : css
     } catch (error) {
       if (error instanceof CssSyntaxError) {
-        logger.error('sheet:toCss', error.message)
-        error.plugin && logger.error('sheet:toCss', `By plugin: ${error.plugin}:`)
-
-        if (error.source) {
-          logger.error('sheet:toCss', `Line ${error.line}:${error.column}, in:`)
-          logger.error('sheet:toCss', error.source)
-        }
+        logger.error('sheet:toCss', error.showSourceCode(true))
       }
 
       throw error