+++ CHANGELOG.md
@@ -6,6 +6,85 @@ See the [Changesets](./.changeset) for the latest changes.
 
 ## [Unreleased]
 
+## [0.25.0] - 2024-01-06
+
+### Fixed
+
+- Fix config dependencies detection by re-introducing the file tracing utility
+- Fix issue where `base` doesn't work within css function
+
+```jsx
+css({
+  // This didn't work, but now it does
+  base: { color: 'blue' },
+})
+```
+
+## Added
+
+- Add a way to generate the staticCss for _all_ recipes (and all variants of each recipe)
+
+```js
+import { defineConfig } from '@pandacss/dev'
+
+export default defineConfig({
+  // ...
+  staticCss: {
+    recipes: '*', // ✅ will generate the staticCss for all recipes
+  },
+})
+```
+
+- Support token reference syntax when authoring styles object, text styles and layer styles.
+
+```jsx
+import { css } from '../styled-system/css'
+
+const styles = css({
+  border: '2px solid {colors.primary}',
+})
+```
+
+This will resolve the token reference and convert it to css variables.
+
+```css
+.border_2px_solid_\{colors\.primary\} {
+  border: 2px solid var(--colors-primary);
+}
+```
+
+The alternative to this was to use the `token(...)` css function which will be resolved.
+
+### `token(...)` vs `{...}`
+
+Both approaches return the css variable
+
+```jsx
+const styles = css({
+  // token reference syntax
+  border: '2px solid {colors.primary}',
+  // token function syntax
+  border: '2px solid token(colors.primary)',
+})
+```
+
+However, The `token(...)` syntax allows you to set a fallback value.
+
+```jsx
+const styles = css({
+  border: '2px solid token(colors.primary, red)',
+})
+```
+
+## [0.24.2] - 2024-01-04
+
+### Fixed
+
+- Fix a regression with utility where boolean values would be treated as a string, resulting in "false" being seen as a
+  truthy value
+- Fix an issue with the `panda init` command which didn't update existing `.gitignore` to include the `styled-system`
+- Fix issue where config slot recipes with compound variants were not processed correctly
+
 ## [0.24.1] - 2024-01-02
 
 ### Fixed