+++ packages/postcss/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @pandacss/postcss
 
+## 0.25.0
+
+### Patch Changes
+
+- Updated dependencies [bc154358]
+  - @pandacss/node@0.25.0
+
+## 0.24.2
+
+### Patch Changes
+
+- @pandacss/node@0.24.2
+
 ## 0.24.1
 
 ### Patch Changes